방 나가기 기능을 위한 모달 컴포넌트 분리 및 스타일 개선
